+++ tests/unit/poller/controller/sources/netbox/test_netbox.py
@@ -14,8 +14,8 @@ from tests.unit.poller.controller.sources.netbox.netbox_faker import \
     NetboxFaker
 from tests.unit.poller.controller.sources.netbox.netbox_rest_server import \
     NetboxRestApp
-from tests.unit.poller.shared.utils import (get_free_port,
-                                            get_src_sample_config)
+from tests.unit.poller.shared.utils import (get_src_sample_config)
+from tests.conftest import get_free_port
 
 # pylint: disable=redefined-outer-name
